.github/requirements.txt
qdrant-client==1.16.1
requests>=2.32.5
kubernetes==35.0.0
GitPython==3.1.46